LPS-86406 Separate ServiceMonitoringControl from ServiceMonitorAdvice and only register once ServiceMonitoringControl in portal. Logically we only had one before since all of the state is static.
