+++ portal-impl/src/portal.properties
@@ -7588,7 +7588,6 @@
     module.framework.services.ignored.interfaces=\
         com.liferay.portal.dao.orm.hibernate.SessionFactoryImpl,\
         com.liferay.portal.kernel.dao.orm.SessionFactory,\
-        com.liferay.portal.monitoring.statistics.service.ServiceMonitorAdvice,\
         com.liferay.portal.service.persistence.impl.*,\
         com.liferay.portal.verify.model.*,\
         com.mchange.*,\
